pyproject.toml
[tool.poetry]
name = "ml-dl-cnn-practice"
version = "0.1.0"
description = "Practice ml, dl, cnn and ai technologies"
authors = ["Ashraful Fuad <fuadashraful16@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ipykernel = "^6.29.4"
pandas = "^2.2.1"
yfinance = "^0.2.37"
scikit-learn = "^1.4.1.post1"
langchain = "^0.3.3"
langchain-openai = "^0.2.2"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
